adding exceptions

+++ src/modules/users/service/index.ts
@@ -3,7 +3,7 @@ import { CreateUserDto } from '../dto';
 import { User } from '../schemas';
 import { UsersRepository } from '../repository';
 import { IUsersService } from '../interfaces';
-
+import { AlreadyExists, NotFound } from 'src/utils/exceptions';
 @Injectable() // permite que a classe seja injetada em outros lugares
 /** @see (Document){IUsersService} */
 export class UsersService implements IUsersService {
@@ -15,6 +15,9 @@ export class UsersService implements IUsersService {
     lastLoginList,
     photo,
   }: CreateUserDto): Promise<User> {
+    if (await this.usersRepository.findOne({ _id: uuid }))
+      throw new AlreadyExists('User');
+
     //cria um usuário no banco de dados e retorna o usuário criado
     return this.usersRepository.create({
       _id: uuid,
@@ -25,14 +28,35 @@ export class UsersService implements IUsersService {
   }
 
   async getUser(id: string): Promise<User> {
-    return this.usersRepository.findOne({ _id: id }); // busca um usuário com o id passado no banco de dados
+    // busca um usuário com o id passado no banco de dados
+    const user = await this.usersRepository.findOne({ _id: id });
+
+    // se não encontrar o usuário, retorna um erro
+    if (!user) throw new NotFound('User');
+
+    // retorna o usuário encontrado
+    return user;
   }
 
   async updateUser(body: User): Promise<User> {
-    return this.usersRepository.findOneAndUpdate({ _id: body._id }, body); // atualiza um usuário no banco de dados
+    // busca um usuário com o id passado no banco de dados
+    const user = await this.usersRepository.findOne({ _id: body._id });
+
+    // se não encontrar o usuário, retorna um erro
+    if (!user) throw new NotFound('User');
+
+    // atualiza o usuário no banco de dados
+    return this.usersRepository.findOneAndUpdate({ _id: body._id }, body);
   }
 
   async deleteUser(id: string): Promise<User> {
+    // busca um usuário com o id passado no banco de dados
+    const user = await this.usersRepository.findOne({ _id: id });
+
+    // se não encontrar o usuário, retorna um erro
+    if (!user) throw new NotFound('User');
+
+    // deleta o usuário no banco de dados
     return this.usersRepository.findOneAndDelete({ _id: id });
   }
 }
